moved several classes to different packages
