+++ src/main/java/apple/excursion/database/VerifyDB.java
@@ -1,6 +1,7 @@
 package apple.excursion.database;
 
 import apple.excursion.ExcursionMain;
+import apple.excursion.database.queries.GetSql;
 import apple.excursion.discord.data.Task;
 import apple.excursion.sheets.SheetsTasks;
