+++ src/main/java/apple/excursion/database/queries/GetSql.java
@@ -1,5 +1,6 @@
-package apple.excursion.database;
+package apple.excursion.database.queries;
 
+import apple.excursion.database.VerifyDB;
 import apple.excursion.discord.data.Task;
 import apple.excursion.discord.data.answers.SubmissionData;
 import apple.excursion.utils.Pair;
