+++ src/main/java/apple/excursion/database/queries/UpdateDB.java
@@ -1,5 +1,6 @@
-package apple.excursion.database;
+package apple.excursion.database.queries;
 
+import apple.excursion.database.VerifyDB;
 import apple.excursion.sheets.SheetsPlayerStats;
 import apple.excursion.utils.Pair;
