+++ src/main/java/apple/excursion/discord/commands/general/leaderboard/CommandCalendar.java
@@ -1,7 +1,7 @@
-package apple.excursion.discord.commands.general;
+package apple.excursion.discord.commands.general.leaderboard;
 
 import apple.excursion.discord.commands.DoCommand;
-import apple.excursion.discord.reactions.messages.CalendarMessage;
+import apple.excursion.discord.reactions.messages.leaderboard.CalendarMessage;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 public class CommandCalendar implements DoCommand {
