+++ src/main/java/apple/excursion/discord/commands/general/leaderboard/CommandLeaderboard.java
@@ -1,7 +1,7 @@
-package apple.excursion.discord.commands.general;
+package apple.excursion.discord.commands.general.leaderboard;
 
 import apple.excursion.discord.commands.DoCommand;
-import apple.excursion.discord.reactions.messages.LeaderboardMessage;
+import apple.excursion.discord.reactions.messages.leaderboard.LeaderboardMessage;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 import java.sql.SQLException;
