+++ src/main/java/apple/excursion/discord/commands/general/self/CommandProfile.java
@@ -1,11 +1,11 @@
-package apple.excursion.discord.commands.general;
+package apple.excursion.discord.commands.general.self;
 
-import apple.excursion.database.GetDB;
+import apple.excursion.database.queries.GetDB;
 import apple.excursion.database.objects.guild.GuildLeaderboardEntry;
 import apple.excursion.database.objects.player.PlayerData;
 import apple.excursion.database.objects.player.PlayerLeaderboardEntry;
 import apple.excursion.discord.commands.DoCommand;
-import apple.excursion.discord.reactions.messages.ProfileMessage;
+import apple.excursion.discord.reactions.messages.self.ProfileMessage;
 import apple.excursion.utils.Pair;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
