+++ src/main/java/apple/excursion/discord/reactions/messages/leaderboard/LeaderboardMessage.java
@@ -1,7 +1,8 @@
-package apple.excursion.discord.reactions.messages;
+package apple.excursion.discord.reactions.messages.leaderboard;
 
-import apple.excursion.database.objects.guild.LeaderboardOfGuilds;
-import apple.excursion.database.objects.guild.GuildLeaderboardEntry;
+import apple.excursion.database.queries.GetDB;
+import apple.excursion.database.objects.player.PlayerLeaderboardEntry;
+import apple.excursion.database.objects.player.PlayerLeaderboard;
 import apple.excursion.discord.reactions.AllReactables;
 import apple.excursion.discord.reactions.ReactableMessage;
 import net.dv8tion.jda.api.entities.Message;
@@ -12,15 +13,14 @@ import net.dv8tion.jda.api.events.message.react.MessageReactionAddEvent;
 import java.sql.SQLException;
 
 
-public class GuildLeaderboardMessage implements ReactableMessage {
-    public static final int ENTRIES_PER_PAGE = 10;
+public class LeaderboardMessage implements ReactableMessage {
+    private final PlayerLeaderboard leaderboard = GetDB.getPlayerLeaderboard();
+    public static final int ENTRIES_PER_PAGE = 20;
     private final Message message;
     private int page;
     private long lastUpdated;
-    private final LeaderboardOfGuilds leaderboard;
 
-    public GuildLeaderboardMessage(MessageChannel channel, LeaderboardOfGuilds leaderboard) throws SQLException {
-        this.leaderboard = leaderboard;
+    public LeaderboardMessage(MessageChannel channel) throws SQLException {
         this.page = 0;
         this.message = channel.sendMessage(makeMessage()).complete();
         message.addReaction(AllReactables.Reactable.LEFT.getFirstEmoji()).queue();
@@ -30,36 +30,30 @@ public class GuildLeaderboardMessage implements ReactableMessage {
         AllReactables.add(this);
     }
 
-    public String makeMessage() {
-        String title = String.format("Excursion Guild Leaderboards Page (%d)", page + 1);
-        return makeMessageStatic(leaderboard,page,title);
+    private String makeMessage() {
+        String title = String.format("Excursion Leaderboards Page (%d)", page + 1);
+        return makeMessageStatic(leaderboard, page, title);
     }
 
-    public static String makeMessageStatic(LeaderboardOfGuilds leaderboard, int page, String title) {
+    public static String makeMessageStatic(PlayerLeaderboard leaderboard, int page, String title) {
         StringBuilder leaderboardMessage = new StringBuilder();
         leaderboardMessage.append(String.format("```glsl\n%s\n", title));
         leaderboardMessage.append(getDash());
         leaderboardMessage.append(String.format("|%4s|", ""));
-        leaderboardMessage.append(String.format("%-20s|", "Guild Name"));
-        leaderboardMessage.append(String.format("%-3s|", "Tag"));
-        leaderboardMessage.append(String.format(" %9s |", "Guild EP"));
-        leaderboardMessage.append(String.format(" %-25s|", "Top Player"));
-        leaderboardMessage.append(String.format(" %-9s|\n", "EP"));
-
+        leaderboardMessage.append(String.format(" %-31s|", "Name"));
+        leaderboardMessage.append(String.format(" %8s |", "Total EP"));
+        leaderboardMessage.append(String.format(" %-20s|", "Guild Name"));
+        leaderboardMessage.append(String.format(" %4s|\n", "Tag "));
         int entriesLength = leaderboard.size();
         for (int place = page * ENTRIES_PER_PAGE; place < ((page + 1) * ENTRIES_PER_PAGE) && place < entriesLength; place++) {
             StringBuilder stringToAdd = new StringBuilder();
             if (place % 5 == 0) {
                 stringToAdd.append(getDash());
             }
-            GuildLeaderboardEntry entry = leaderboard.get(place);
-            stringToAdd.append(String.format("|%4d|%-20s|%-3s| %9d | %-25s| %-9d|\n",
-                    entry.rank,
-                    entry.getGuildName(),
-                    entry.getGuildTag(),
-                    entry.score,
-                    entry.topPlayer.length() > 25 ? entry.topPlayer.substring(0, 22) + "..." : entry.topPlayer,
-                    entry.topPlayerPoints));
+            PlayerLeaderboardEntry entry = leaderboard.get(place);
+            final String name = entry.playerName;
+            stringToAdd.append(String.format("|%4d| %-31s| %8d | %-20s| %3s |\n",
+                    place + 1, name.length() > 25 ? name.substring(0, 22) + "..." : name, entry.score, entry.getGuildName(), entry.getGuildTag()));
 
             if (leaderboardMessage.length() + 3 + stringToAdd.length() >= 2000) {
                 leaderboardMessage.append("```");
@@ -68,10 +62,6 @@ public class GuildLeaderboardMessage implements ReactableMessage {
                 leaderboardMessage.append(stringToAdd);
             }
         }
-        leaderboardMessage.append(getDash());
-        leaderboardMessage.append(String.format("Total EP: %d EP\n", leaderboard.getTotalEp()));
-        leaderboardMessage.append(String.format("Total EP guildless submissions: %d EP\n", leaderboard.getNoGuildsEp()));
-        leaderboardMessage.append(getDash());
         leaderboardMessage.append("```");
         return leaderboardMessage.toString();
     }
