+++ src/main/java/apple/excursion/discord/reactions/messages/self/PostcardListMessage.java
@@ -1,4 +1,4 @@
-package apple.excursion.discord.reactions.messages;
+package apple.excursion.discord.reactions.messages.self;
 
 import apple.excursion.discord.DiscordBot;
 import apple.excursion.discord.data.Task;
